refactor side, making presentational logic clearer
