+++ ui/coordinateTrainer/src/view.ts
@@ -2,194 +2,8 @@ import { h, VNode, VNodeStyle } from 'snabbdom';
 import { bind, MaybeVNode } from 'common/snabbdom';
 import chessground from './chessground';
 import CoordinateTrainerCtrl, { DURATION } from './ctrl';
-import { ColorChoice, TimeControl, Mode, CoordModifier } from './interfaces';
-
-const scoreCharts = (ctrl: CoordinateTrainerCtrl): VNode => {
-  const average = (array: number[]) => array.reduce((a, b) => a + b) / array.length;
-  return h(
-    'div.scores',
-    [
-      ['sente', 'averageScoreAsWhiteX', ctrl.modeScores[ctrl.mode].sente],
-      ['gote', 'averageScoreAsBlackX', ctrl.modeScores[ctrl.mode].gote],
-    ].map(([color, transKey, scoreList]: [Color, string, number[]]) =>
-      scoreList.length
-        ? h('div.color-chart', [
-            h('p', ctrl.trans.vdom(transKey, h('strong', `${average(scoreList).toFixed(2)}`))),
-            h('svg.sparkline', {
-              attrs: {
-                height: '80px',
-                'stroke-width': '3',
-                id: `${color}-sparkline`,
-              },
-              hook: {
-                insert: vnode => ctrl.updateChart(vnode.elm as SVGSVGElement, color),
-              },
-            }),
-          ])
-        : null
-    )
-  );
-};
-
-const colors: [ColorChoice, string][] = [
-  ['black', 'asBlack'],
-  ['random', 'randomColor'],
-  ['white', 'asWhite'],
-];
-
-const timeControls: [TimeControl, string][] = [
-  ['untimed', '∞'],
-  ['thirtySeconds', '0:30'],
-];
-
-const side = (ctrl: CoordinateTrainerCtrl): VNode => {
-  const { trans } = ctrl;
-
-  const sideContent: MaybeVNode[] = [h('div.box', h('h1', trans('coordinates')))];
-  if (ctrl.playing || ctrl.hasPlayed) {
-    sideContent.push(
-      h('div.box.current-status', [h('h1', trans('score')), h('div.score', ctrl.score)]),
-      ctrl.playing && !ctrl.timeDisabled()
-        ? h('div.box.current-status', [
-            h('h1', trans('time')),
-            h('div.timer', { class: { hurry: ctrl.timeLeft <= 10 * 1000 } }, (ctrl.timeLeft / 1000).toFixed(1)),
-          ])
-        : null
-    );
-  }
-  if (!ctrl.playing) {
-    sideContent.push(
-      h('form.mode.buttons', [
-        h(
-          'group.radio',
-          ['findSquare', 'nameSquare'].map((mode: Mode) =>
-            h('div.mode_option', [
-              h('input', {
-                attrs: {
-                  type: 'radio',
-                  id: `coord_mode_${mode}`,
-                  name: 'mode',
-                  value: mode,
-                  checked: mode === ctrl.mode,
-                },
-                on: {
-                  change: e => {
-                    const target = e.target as HTMLInputElement;
-                    ctrl.setMode(target.value as Mode);
-                  },
-                  keyup: ctrl.onRadioInputKeyUp,
-                },
-              }),
-              h(`label.mode_${mode}`, { attrs: { for: `coord_mode_${mode}` } }, trans(mode)),
-            ])
-          )
-        ),
-      ]),
-      h('form.timeControl.buttons', [
-        h(
-          'group.radio',
-          timeControls.map(([timeControl, timeControlLabel]) =>
-            h('div.timeControl_option', [
-              h('input', {
-                attrs: {
-                  type: 'radio',
-                  id: `coord_timeControl_${timeControl}`,
-                  name: 'timeControl',
-                  value: timeControl,
-                  checked: timeControl === ctrl.timeControl,
-                },
-                on: {
-                  change: e => {
-                    const target = e.target as HTMLInputElement;
-                    ctrl.setTimeControl(target.value as TimeControl);
-                  },
-                  keyup: ctrl.onRadioInputKeyUp,
-                },
-              }),
-              h(
-                `label.timeControl_${timeControl}`,
-                {
-                  attrs: {
-                    for: `coord_timeControl_${timeControl}`,
-                  },
-                },
-                timeControlLabel
-              ),
-            ])
-          )
-        ),
-      ]),
-      h('form.color.buttons', [
-        h(
-          'group.radio',
-          ['gote', 'random', 'sente'].map((color: ColorChoice) =>
-            h('div', [
-              h('input', {
-                attrs: {
-                  type: 'radio',
-                  id: `coord_color_${color}`,
-                  name: 'color',
-                  value: color,
-                  checked: color === ctrl.colorChoice,
-                },
-                on: {
-                  change: e => {
-                    const target = e.target as HTMLInputElement;
-                    ctrl.setColorChoice(target.value as ColorChoice);
-                  },
-                  keyup: ctrl.onRadioInputKeyUp,
-                },
-              }),
-              h(
-                `label.color_${color}`,
-                {
-                  attrs: {
-                    type: 'radio',
-                    id: `coord_color_${color}`,
-                    name: 'color',
-                    value: color,
-                    checked: color === ctrl.colorChoice,
-                  },
-                  on: {
-                    change: e => {
-                      const target = e.target as HTMLInputElement;
-                      ctrl.setColorChoice(target.value as ColorChoice);
-                    },
-                  },
-                }),
-                h(
-                  `label.color_${color}`,
-                  {
-                    attrs: {
-                      for: `coord_color_${color}`,
-                    },
-                  },
-                  h('i')
-                ),
-              ])
-            )
-          ),
-        ]),
-      ]
-    );
-  }
-  if (ctrl.isAuth && ctrl.hasModeScores()) sideContent.push(h('div.box', scoreCharts(ctrl)));
-  if (ctrl.playing && ctrl.timeDisabled())
-    sideContent.push(
-      h(
-        'div.back',
-        h(
-          'a.back-button',
-          {
-            hook: bind('click', ctrl.stop),
-          },
-          `« ${trans('back')}`
-        )
-      )
-    );
-
-  return h('div.side', sideContent);
-};
+import { CoordModifier } from './interfaces';
+import side from './side';
 
 const board = (ctrl: CoordinateTrainerCtrl): VNode => {
   return h('div.main-board', [
